main/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.2
Django==5.0.1
django-filter==23.5
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.6
jwcrypto==1.5.3
Markdown==3.5.2
mysqlclient==2.2.3
oauthlib==3.2.2
pillow==10.2.0
pycparser==2.21
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.0
